.vimrc: don't restore gitcommit cursor position
Each time I commit, I have to edit a different commit message, and almost
every time I have to begin writing a completely new message, but the cursor is
not there, because with my configuration Vim restores the old position for the
previous commit message. The file path is the same, after all.

Change the configuration NOT to restore the cursor position for the gitcommit
file type.